python-code/code/requirements.txt
matplotlib==3.0.2
mtcnn==0.1.0
numpy==1.16.0
pillow==5.3.0
scipy==1.1.0
tensorflow==1.12.0
keras==2.2.4
opencv-python==4.1.0.25
flask==1.0.2
keras_vggface==0.6
imutils==0.5.1
pyinstaller==3.4
setuptools==39.0.1
